+++ RackPeek.Domain/Resources/Hardware/Models/Port.cs
@@ -3,6 +3,6 @@ namespace RackPeek.Domain.Resources.Hardware.Models;
 public class Port
 {
     public string? Type { get; set; }
-    public int? Speed { get; set; }
+    public double? Speed { get; set; }
     public int? Count { get; set; }
 }
