+++ RackPeek.Domain/Resources/Hardware/Routers/DescribeRouterUseCase.cs
@@ -9,7 +9,7 @@ public record RouterDescription(
     bool? Managed,
     bool? Poe,
     int TotalPorts,
-    int TotalSpeedGb,
+    double TotalSpeedGb,
     string PortSummary
 );
